requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-prometheus==1.24.0
motor
pydantic[email]
python-jose
passlib[bcrypt]